requirements.txt
click
cachetools
python-chess
stockfish